Move test shouldReturnBadGatewayWhenUnauthorized to BaseCreateSerialPublicationHandlerTest
